// lib/types.ts
export type AOR = { name: string; uri: string; params?: Record<string, any>; };
export type Via = { version: string; protocol: string; host: string; port: string; };

export interface SrfConfig {
  apiSecret?: string;
  host?: string;
  port?: number;
  secret?: string;
  tags?: string[];
}

export interface ParseUriResult {
  family?: 'ipv6' | 'ipv4';
  scheme: 'sip' | 'sips' | 'tel';
  user?: string;
  password?: string;
  host?: string;
  port?: string;
  params: Record<string, string | null>;
  headers: Record<string, string>;
  context?: string;
}

export interface SipMessage {
  type: "request" | "response";
  body: string;
  payload: object[];
  source: "network" | "application";
  source_address: string;
  source_port: string;
  protocol: string;
  stackTime: string;
  calledNumber: string;
  callingNumber: string;
  raw: string;
  get(name: string): string;
  has(name: string): boolean;
  set(name: string, value: string): void;
  getParsedHeader(name: "contact" | "Contact"): Array<AOR>;
  getParsedHeader(name: "via" | "Via"): Array<Via>;
  getParsedHeader(name: "To" | "to" | "From" | "from" | "refer-to" | "referred-by" | "p-asserted-identity" | "remote-party-id"): AOR;
  getParsedHeader(name: string): any;
}